FriendAvatar now split up into Avatar, FriendAvatar and LobbyAvatar
