+++ Assets/Prefabs/LobbyAvatar.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d1a5d96118289b42a25fe5890eb98d7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
