+++ Assets/Scripts/Networking/Avatar.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8b77f55d437675a41badd46bc14f8804
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
