+++ Assets/Scripts/Networking/Avatar/Avatar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b38f037832c8ac94c884218bf091a0c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
