+++ Assets/Scripts/Networking/Avatar/LobbyAvatar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e46ecdd4fe7565d40bce07e9bffffc46
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
